refactor(dataset): move `withDataPrefix` to `datasetKeys` test file
